backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.12
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.79
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.1
ultralytics>=8.0.0
pyyaml>=6.0
